requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
mysql-connector-python==8.2.0
Werkzeug==3.0.1
python-dotenv==1.2.2
waitress==3.0.2
Jinja2==3.1.6
